Updating async_pkey_test to use source file; Adding proof for async handler call
