+++ tests/unit/s2n_async_pkey_test.c
@@ -26,11 +26,12 @@
 #include "utils/s2n_safety.h"
 
 /* Include source to test on a specific pkey_op type */
-#include "tls/s2n_async_pkey.h"
+#include "tls/s2n_async_pkey.c"
 
 struct s2n_async_pkey_op *pkey_op = NULL;
 
 typedef int (async_handler)(struct s2n_connection *conn);
+static int async_handler_called = 0;
 
 static int async_handler_fail(struct s2n_connection *conn)
 {
@@ -101,9 +102,15 @@ static int async_handler_sign_with_different_pkey_and_apply(struct s2n_connectio
         /* Set chain_and_key back to original value and free new chain_and_key */
         conn->handshake_params.our_chain_and_key = chain_and_key;
         EXPECT_SUCCESS(s2n_cert_chain_and_key_free(chain_and_key_2));
+
+        /* Increment flag counter for sign operation */
+        async_handler_called++;
     } else {
         /* Test decrypt operation passes */
         EXPECT_SUCCESS(s2n_async_pkey_op_apply(pkey_op, conn));
+
+        /* Increment flag counter for decrypt operation */
+        async_handler_called++;
     }
 
     /* Free the pkey op */
@@ -379,6 +386,10 @@ int main(int argc, char **argv)
             EXPECT_SUCCESS(s2n_connection_set_io_pair(server_conn, &io_pair));
 
             EXPECT_SUCCESS(try_handshake(server_conn, client_conn, async_handler_sign_with_different_pkey_and_apply));
+            EXPECT_EQUAL(async_handler_called, 1);
+
+            /* Reset counter */
+            async_handler_called = 0;
 
             /* Free the data */
             EXPECT_SUCCESS(s2n_connection_free(server_conn));
